Style nav elements with CSS grid
Instead of using nested flexboxes, if we break out the two `<ul>`s to be sibling elements of `<form id="p-search">` and style them all with a CSS grid and a new breakpoint, then when the Actions menu wraps to a layer above the others the search bar can tuck neatly underneath it. Voila!