Update `main.css` from SVN `r1596`
From SVN log `r1596`:
> undo r1583 because i see now i had a smart reason for it originally, which i apparently forgot ... the raster background texture gets cuts off unless we tell the browser that the border box of the div it's tied to continues to the bottom of the page. we'll have to figure out some other solution for the wiki pages

To handle this on the Wiki skin side, I played with adding classes to the DOM elements in `skin.mustache`.

I was prepared to add both `copy-main` and `copy-main-last` so we would actually have the cool raster texture in the Wiki article body ... It was occluding the top level menu (which it doesn't on the main site due to presence of the theatre element), which was going to be a pretty easy fix ... But looking at it, I ultimately made an artistic decision that the Wiki pages look better without the vellum in the center.

So we are only adding the `copy-main-last` class. Problem solved!